+++ toolium/driver_wrappers_pool.py
@@ -251,11 +251,10 @@ class DriverWrappersPool(object):
             driver_index += 1
 
     @staticmethod
-    def get_configured_value(system_property_name, deprecated_system_property_name, specific_value, default_value):
+    def get_configured_value(system_property_name, specific_value, default_value):
         """Get configured value from system properties, method parameters or default value
 
         :param system_property_name: system property name
-        :param deprecated_system_property_name: deprecated system property name
         :param specific_value: test case specific value
         :param default_value: default value
         :returns: configured value
@@ -263,10 +262,7 @@ class DriverWrappersPool(object):
         try:
             return os.environ[system_property_name]
         except KeyError:
-            try:
-                return os.environ[deprecated_system_property_name]
-            except KeyError:
-                return specific_value if specific_value else default_value
+            return specific_value if specific_value else default_value
 
     @classmethod
     def configure_common_directories(cls, tc_config_files):
@@ -276,14 +272,14 @@ class DriverWrappersPool(object):
         """
         if cls.config_directory is None:
             # Get config directory from properties
-            config_directory = cls.get_configured_value('TOOLIUM_CONFIG_DIRECTORY', 'Config_directory',
+            config_directory = cls.get_configured_value('TOOLIUM_CONFIG_DIRECTORY',
                                                         tc_config_files.config_directory, 'conf')
-            prop_filenames = cls.get_configured_value('TOOLIUM_CONFIG_PROPERTIES_FILENAMES', 'Config_prop_filenames',
+            prop_filenames = cls.get_configured_value('TOOLIUM_CONFIG_PROPERTIES_FILENAMES',
                                                       tc_config_files.config_properties_filenames, 'properties.cfg')
             cls.config_directory = cls._find_parent_directory(config_directory, prop_filenames.split(';')[0])
 
             # Get output directory from properties and create it
-            cls.output_directory = cls.get_configured_value('TOOLIUM_OUTPUT_DIRECTORY', 'Output_directory',
+            cls.output_directory = cls.get_configured_value('TOOLIUM_OUTPUT_DIRECTORY',
                                                             tc_config_files.output_directory, 'output')
             if not os.path.isabs(cls.output_directory):
                 # If output directory is relative, we use the same path as config directory
@@ -293,7 +289,6 @@ class DriverWrappersPool(object):
             # Get visual baseline directory from properties
             default_baseline = os.path.join(cls.output_directory, 'visualtests', 'baseline')
             cls.visual_baseline_directory = cls.get_configured_value('TOOLIUM_VISUAL_BASELINE_DIRECTORY',
-                                                                     'Visual_baseline_directory',
                                                                      tc_config_files.visual_baseline_directory,
                                                                      default_baseline)
             if not os.path.isabs(cls.visual_baseline_directory):
@@ -364,7 +359,7 @@ class DriverWrappersPool(object):
             tc_config_files = ConfigFiles()
 
         # Update properties and log file names if an environment is configured
-        env = DriverWrappersPool.get_configured_value('TOOLIUM_CONFIG_ENVIRONMENT', 'Config_environment', None, None)
+        env = DriverWrappersPool.get_configured_value('TOOLIUM_CONFIG_ENVIRONMENT', None, None)
         if env:
             # Update config properties filenames
             prop_filenames = tc_config_files.config_properties_filenames